refactor: move fungible assets tests to seperate file
