+++ tests/xcm-simulator/src/lib.rs
@@ -20,6 +20,4 @@
 mod mocks;
 
 #[cfg(test)]
-mod experimental;
-#[cfg(test)]
 mod tests;
