+++ tests/xcm-simulator/src/tests/general.rs
@@ -128,433 +128,6 @@ fn basic_xcmp() {
 }
 
 #[test]
-fn para_to_para_reserve_transfer_and_back() {
-    MockNet::reset();
-
-    let sibling_asset_id = 123 as u128;
-    let para_a_multiloc = (Parent, Parachain(1));
-
-    // On parachain B create an asset which representes a derivative of parachain A native asset.
-    // This asset is allowed as XCM execution fee payment asset.
-    ParaB::execute_with(|| {
-        assert_ok!(register_and_setup_xcm_asset::<parachain::Runtime, _>(
-            parachain::RuntimeOrigin::root(),
-            sibling_asset_id,
-            para_a_multiloc.clone(),
-            sibling_para_account_id(1),
-            Some(true),
-            Some(1),
-            Some(1_000_000_000_000)
-        ));
-    });
-
-    // Next step is to send some of parachain A native asset to parachain B.
-    let withdraw_amount = 567;
-    ParaA::execute_with(|| {
-        assert_ok!(ParachainPalletXcm::reserve_transfer_assets(
-            parachain::RuntimeOrigin::signed(ALICE),
-            Box::new(MultiLocation::new(1, X1(Parachain(2))).into()),
-            Box::new(
-                X1(AccountId32 {
-                    network: None,
-                    id: ALICE.into()
-                })
-                .into_location()
-                .into_versioned()
-            ),
-            Box::new((Here, withdraw_amount).into()),
-            0,
-        ));
-
-        // Parachain 2 sovereign account should have it's balance increased, while Alice balance should be decreased.
-        assert_eq!(
-            parachain::Balances::free_balance(&sibling_para_account_id(2)),
-            INITIAL_BALANCE + withdraw_amount
-        );
-        assert_eq!(
-            parachain::Balances::free_balance(&ALICE),
-            INITIAL_BALANCE - withdraw_amount
-        );
-    });
-
-    // Parachain B should receive parachain A native assets and should mint their local derivate.
-    // Portion of those assets should be taken as the XCM execution fee.
-    let four_instructions_execution_cost =
-        (parachain::UnitWeightCost::get() * 4).ref_time() as u128;
-    let remaining = withdraw_amount - four_instructions_execution_cost;
-    ParaB::execute_with(|| {
-        // Ensure Alice received assets on ParaB (sent amount minus expenses)
-        assert_eq!(
-            parachain::Assets::balance(sibling_asset_id, ALICE),
-            remaining
-        );
-    });
-
-    // send assets back to ParaA
-    ParaB::execute_with(|| {
-        assert_ok!(ParachainPalletXcm::reserve_withdraw_assets(
-            parachain::RuntimeOrigin::signed(ALICE),
-            Box::new((Parent, Parachain(1)).into()),
-            Box::new(
-                AccountId32 {
-                    network: None,
-                    id: ALICE.into()
-                }
-                .into()
-            ),
-            Box::new((para_a_multiloc, remaining).into()),
-            0
-        ));
-    });
-
-    ParaA::execute_with(|| {
-        // ParaB soveregin account account should have only the execution cost
-        assert_eq!(
-            parachain::Balances::free_balance(&sibling_para_account_id(2)),
-            INITIAL_BALANCE + four_instructions_execution_cost
-        );
-        // ParaA alice should have initial amount backed subtracted with execution costs
-        // which is 2xfour_instructions_execution_cost
-        // or withdraw_amount + remaining - four_instructions_execution_cost
-        // both are same
-        assert_eq!(
-            parachain::Balances::free_balance(&ALICE),
-            INITIAL_BALANCE - withdraw_amount + remaining - four_instructions_execution_cost
-        );
-    });
-}
-
-#[test]
-fn para_to_para_reserve_transfer_local_asset() {
-    MockNet::reset();
-
-    let asset_id = 123;
-    let local_asset: MultiLocation = (PalletInstance(4u8), GeneralIndex(asset_id)).into();
-    let para_a_local_asset = local_asset
-        .clone()
-        .pushed_front_with_interior(Parachain(1))
-        .unwrap()
-        .prepended_with(Parent)
-        .unwrap();
-
-    ParaA::execute_with(|| {
-        assert_ok!(register_and_setup_xcm_asset::<parachain::Runtime, _>(
-            parachain::RuntimeOrigin::root(),
-            asset_id,
-            local_asset,
-            ALICE.into(),
-            Some(true),
-            Some(1),
-            // free execution
-            Some(0)
-        ));
-
-        assert_ok!(ParachainAssets::mint(
-            parachain::RuntimeOrigin::signed(ALICE.into()),
-            asset_id,
-            ALICE.into(),
-            300000000000000
-        ));
-    });
-
-    ParaB::execute_with(|| {
-        assert_ok!(register_and_setup_xcm_asset::<parachain::Runtime, _>(
-            parachain::RuntimeOrigin::root(),
-            asset_id,
-            para_a_local_asset,
-            sibling_para_account_id(1),
-            Some(true),
-            Some(1),
-            // free execution
-            Some(0)
-        ));
-    });
-
-    let send_amount = 123;
-    ParaA::execute_with(|| {
-        assert_ok!(ParachainPalletXcm::reserve_transfer_assets(
-            parachain::RuntimeOrigin::signed(ALICE.into()),
-            Box::new((Parent, Parachain(2)).into()),
-            Box::new(
-                AccountId32 {
-                    network: None,
-                    id: ALICE.into(),
-                }
-                .into(),
-            ),
-            Box::new((local_asset, send_amount).into()),
-            0,
-        ));
-    });
-
-    ParaB::execute_with(|| {
-        // free execution, full amount received
-        assert_eq!(
-            ParachainAssets::balance(asset_id, &ALICE.into()),
-            send_amount
-        );
-    });
-}
-// Send a relay asset (like DOT/KSM) to a parachain A
-#[test]
-fn receive_relay_asset_from_relay() {
-    MockNet::reset();
-
-    let relay_asset_id = 123 as u128;
-    let source_location = (Parent,);
-
-    // On parachain A create an asset which representes a derivative of relay native asset.
-    // This asset is allowed as XCM execution fee payment asset.
-    ParaA::execute_with(|| {
-        assert_ok!(register_and_setup_xcm_asset::<parachain::Runtime, _>(
-            parachain::RuntimeOrigin::root(),
-            relay_asset_id,
-            source_location,
-            parent_account_id(),
-            Some(true),
-            Some(1),
-            Some(1_000_000_000_000)
-        ));
-    });
-
-    // Next step is to send some of relay native asset to parachain A.
-    let withdraw_amount = 567;
-    Relay::execute_with(|| {
-        assert_ok!(RelayChainPalletXcm::reserve_transfer_assets(
-            relay_chain::RuntimeOrigin::signed(ALICE),
-            Box::new(Parachain(1).into()),
-            Box::new(
-                AccountId32 {
-                    network: None,
-                    id: ALICE.into()
-                }
-                .into_location()
-                .into_versioned()
-            ),
-            Box::new((Here, withdraw_amount).into()),
-            0,
-        ));
-
-        // Parachain A sovereign account should have it's balance increased, while Alice balance should be decreased.
-        assert_eq!(
-            relay_chain::Balances::free_balance(&child_para_account_id(1)),
-            INITIAL_BALANCE + withdraw_amount
-        );
-        assert_eq!(
-            relay_chain::Balances::free_balance(&ALICE),
-            INITIAL_BALANCE - withdraw_amount
-        );
-    });
-
-    // Parachain A should receive relay native assets and should mint their local derivate.
-    // Portion of those assets should be taken as the XCM execution fee.
-    ParaA::execute_with(|| {
-        // Ensure Alice received assets on ParaA (sent amount minus expenses)
-        let four_instructions_execution_cost = parachain::UnitWeightCost::get() * 4;
-        assert_eq!(
-            parachain::Assets::balance(relay_asset_id, ALICE),
-            withdraw_amount - four_instructions_execution_cost.ref_time() as u128
-        );
-    });
-}
-
-// Send relay asset (like DOT) back from Parachain A to relaychain
-#[test]
-fn send_relay_asset_to_relay() {
-    MockNet::reset();
-
-    let source_location = (Parent,);
-    let relay_asset_id = 123_u128;
-    let alice = AccountId32 {
-        network: None,
-        id: ALICE.into(),
-    };
-
-    // On parachain A create an asset which representes a derivative of relay native asset.
-    // This asset is allowed as XCM execution fee payment asset.
-    // Register relay asset in paraA
-    ParaA::execute_with(|| {
-        assert_ok!(register_and_setup_xcm_asset::<parachain::Runtime, _>(
-            parachain::RuntimeOrigin::root(),
-            relay_asset_id,
-            source_location,
-            parent_account_id(),
-            Some(true),
-            Some(123),
-            Some(0)
-        ));
-    });
-
-    // Next step is to send some of relay native asset to parachain A.
-    // same as previous test
-    let withdraw_amount = 54321;
-    Relay::execute_with(|| {
-        assert_ok!(RelayChainPalletXcm::reserve_transfer_assets(
-            relay_chain::RuntimeOrigin::signed(ALICE),
-            Box::new(Parachain(1).into()),
-            Box::new(alice.into_location().into_versioned()),
-            Box::new((Here, withdraw_amount).into()),
-            0,
-        ));
-    });
-
-    ParaA::execute_with(|| {
-        // Free execution, full amount received
-        assert_eq!(
-            parachain::Assets::balance(relay_asset_id, ALICE),
-            withdraw_amount
-        );
-    });
-
-    // Lets gather the balance before sending back money
-    let mut balance_before_sending = 0;
-    Relay::execute_with(|| {
-        balance_before_sending = relay_chain::Balances::free_balance(&ALICE);
-    });
-
-    ParaA::execute_with(|| {
-        assert_ok!(ParachainPalletXcm::reserve_withdraw_assets(
-            parachain::RuntimeOrigin::signed(ALICE),
-            Box::new(Parent.into()),
-            Box::new(alice.into_location().into_versioned()),
-            Box::new((Parent, withdraw_amount).into()),
-            0,
-        ));
-    });
-
-    // The balances in ParaA alice should have been substracted
-    ParaA::execute_with(|| {
-        assert_eq!(parachain::Assets::balance(relay_asset_id, ALICE), 0);
-    });
-
-    // Balances in the relay should have been received
-    Relay::execute_with(|| {
-        // free execution,x	 full amount received
-        assert!(relay_chain::Balances::free_balance(ALICE) > balance_before_sending);
-    });
-
-    // // To get logs
-    // std::thread::sleep(std::time::Duration::from_millis(4000));
-}
-
-// Send relay asset (like DOT) back from Parachain A to Parachain B
-#[test]
-fn para_a_send_relay_asset_to_para_b() {
-    MockNet::reset();
-
-    let source_location = (Parent,);
-    let relay_asset_id = 123_u128;
-    let alice = AccountId32 {
-        network: None,
-        id: ALICE.into(),
-    };
-
-    // On parachain A create an asset which representes a derivative of relay native asset.
-    // This asset is allowed as XCM execution fee payment asset.
-    // Register relay asset in ParaA
-    ParaA::execute_with(|| {
-        assert_ok!(register_and_setup_xcm_asset::<parachain::Runtime, _>(
-            parachain::RuntimeOrigin::root(),
-            relay_asset_id,
-            source_location,
-            parent_account_id(),
-            Some(true),
-            Some(123),
-            // free execution
-            Some(0)
-        ));
-    });
-
-    // register relay asset in ParaB
-    ParaB::execute_with(|| {
-        assert_ok!(register_and_setup_xcm_asset::<parachain::Runtime, _>(
-            parachain::RuntimeOrigin::root(),
-            relay_asset_id,
-            source_location,
-            parent_account_id(),
-            Some(true),
-            Some(123),
-            Some(1_000_000_000_000)
-        ));
-    });
-
-    // Next step is to send some of relay native asset to parachain A.
-    // same as previous test
-    let withdraw_amount = 54321;
-    Relay::execute_with(|| {
-        assert_ok!(RelayChainPalletXcm::reserve_transfer_assets(
-            relay_chain::RuntimeOrigin::signed(ALICE),
-            Box::new(Parachain(1).into()),
-            Box::new(alice.into_location().into_versioned()),
-            Box::new((Here, withdraw_amount).into()),
-            0,
-        ));
-    });
-
-    ParaA::execute_with(|| {
-        // Free execution, full amount received
-        assert_eq!(
-            parachain::Assets::balance(relay_asset_id, ALICE),
-            withdraw_amount
-        );
-    });
-
-    // send relay asset to ParaB
-    ParaA::execute_with(|| {
-        let xcm = Xcm(vec![
-            // withdraw relay native asset
-            WithdrawAsset((Parent, withdraw_amount).into()),
-            InitiateReserveWithdraw {
-                assets: All.into(),
-                reserve: source_location.clone().into(),
-                xcm: Xcm(vec![
-                    BuyExecution {
-                        fees: (Here, 1).into(),
-                        weight_limit: Unlimited,
-                    },
-                    // deposit into ParaB
-                    DepositReserveAsset {
-                        assets: All.into(),
-                        dest: Parachain(2).into(),
-                        xcm: Xcm(vec![
-                            BuyExecution {
-                                // cost for 4 intr
-                                fees: (Parent, 40).into(),
-                                weight_limit: Unlimited,
-                            },
-                            // deposit into ParaB's alice
-                            DepositAsset {
-                                assets: All.into(),
-                                beneficiary: alice.clone().into(),
-                            },
-                        ]),
-                    },
-                ]),
-            },
-        ]);
-
-        assert_ok!(ParachainPalletXcm::execute(
-            parachain::RuntimeOrigin::signed(ALICE.into()),
-            Box::new(VersionedXcm::V3(xcm)),
-            Weight::from_parts(100_000_000_000, 1024 * 1024)
-        ));
-    });
-
-    // Para A balances should have been substracted
-    ParaA::execute_with(|| {
-        assert_eq!(parachain::Assets::balance(relay_asset_id, ALICE), 0);
-    });
-
-    // Para B balances should have been credited
-    ParaB::execute_with(|| {
-        assert_eq!(
-            parachain::Assets::balance(relay_asset_id, ALICE),
-            withdraw_amount - 40
-        );
-    });
-}
-
-#[test]
 fn receive_asset_with_no_sufficients_not_possible_if_non_existent_account() {
     MockNet::reset();
