+++ tests/xcm-simulator/src/tests/mod.rs
@@ -1,0 +1,8 @@
+#[cfg(test)]
+mod experimental;
+
+#[cfg(test)]
+mod fungible_assets;
+
+#[cfg(test)]
+mod general;
